refactor(shared): sort types and schemas to correct folders
