+++ packages/shared/src/types/game-session.ts
@@ -1,0 +1,4 @@
+import type z from "zod"
+import type { GameSessionScheduleSchema } from "@/schemas"
+
+export type GameSessionSchedule = z.infer<typeof GameSessionScheduleSchema>
