+++ packages/shared/src/types/index.ts
@@ -4,3 +4,4 @@ export * from "./enums"
 export * from "./payload"
 export * from "./richtext"
 export * from "./semester"
+export * from "./user"
