+++ packages/shared/src/types/semester.ts
@@ -1,18 +1,4 @@
-import { z } from "zod"
-import type { CreateSemesterData, EditSemesterData } from "./collections"
-import { Weekday } from "./enums"
+import type { z } from "zod"
+import type { CreateSemesterRequestSchema } from "@/schemas"
 
-export const CreateSemesterRequestSchema = z.object({
-  name: z.string(),
-  startDate: z.string().datetime({ message: "Invalid date format, should be in ISO 8601 format" }),
-  endDate: z.string().datetime({ message: "Invalid date format, should be in ISO 8601 format" }),
-  breakStart: z.string().datetime({ message: "Invalid date format, should be in ISO 8601 format" }),
-  breakEnd: z.string().datetime({ message: "Invalid date format, should be in ISO 8601 format" }),
-  bookingOpenDay: z.nativeEnum(Weekday),
-  bookingOpenTime: z
-    .string()
-    .datetime({ message: "Invalid date format, should be in ISO 8601 format" }),
-}) satisfies z.ZodType<CreateSemesterData>
-
-export const UpdateSemesterRequestSchema =
-  CreateSemesterRequestSchema.partial() satisfies z.ZodType<EditSemesterData>
+export type CreateSemesterRequest = z.infer<typeof CreateSemesterRequestSchema>
