+++ packages/shared/src/types/user.ts
@@ -1,0 +1,5 @@
+import type z from "zod"
+import type { CreateUserRequestSchema, UpdateUserRequestSchema } from "@/schemas"
+
+export type CreateUserRequest = z.infer<typeof CreateUserRequestSchema>
+export type UpdateUserRequest = z.infer<typeof UpdateUserRequestSchema>
